Updates to Angular 20 and Nx 21
Bumps the Angular and Nx versions to the 20.x.x and 21.x.x lines, respectively.

This change includes updates to dependencies in the package.json files, adjustments to project configurations, and minor code changes to align with the updated versions. It also removes a duplicate package.json file.

The updated versions bring performance improvements and updated features.